Add files via upload
It started Bestiary one, as you can see it was a mess. Then I thought I just scrap it and started over before I had something to lose. Bestiary2 is almost like one but I made it cause I realized i needed a database to actually make what I'm doing a lot easier and a lot more cleaner, so I used Postgres to connect to a server like a database and I looked up on how I can use Postgres as a server that I can turn off and on whenever I please. 

Then I realized that it's still going to be a mess cause the code was starting to be more than what's needed. So I remembered about OOP and used it with Tkinter to create easy buttons and windows. it took a little time to figure it out if you looked at the test files. but then I finally got to the last phase which was putting it all together. Now My 
"Software App" can retrieve information from the database whenever a button is clicked for the person to find out more information about the said chosen Shin Megami Tensei Demon. 

Lessons Learned: 
OOP Tkinter
Database Send and Retrieve